refactor: separate applicationCommands files into folders
