+++ payloads/v9/_interactions/_applicationCommands/contextMenu.ts
@@ -1,0 +1,104 @@
+import type { APIUser } from '../../index';
+import type { Snowflake } from '../../../../globals';
+import type { APIMessage } from '../../channel';
+import type { APIDMInteractionWrapper, APIGuildInteractionWrapper } from '../base';
+import type {
+	APIApplicationCommandInteractionWrapper,
+	APIBaseApplicationCommandInteractionData,
+	APIInteractionDataResolvedGuildMember,
+	ApplicationCommandType,
+} from '../applicationCommands';
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-interaction-data-structure
+ */
+export interface APIUserApplicationCommandInteractionData
+	extends APIBaseApplicationCommandInteractionData<ApplicationCommandType.User> {
+	target_id: Snowflake;
+	resolved: APIUserApplicationCommandInteractionDataResolved;
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-resolved-data-structure
+ */
+export interface APIUserApplicationCommandInteractionDataResolved {
+	users: Record<Snowflake, APIUser>;
+	members?: Record<Snowflake, APIInteractionDataResolvedGuildMember>;
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-interaction-data-structure
+ */
+export interface APIMessageApplicationCommandInteractionData
+	extends APIBaseApplicationCommandInteractionData<ApplicationCommandType.Message> {
+	target_id: Snowflake;
+	resolved: APIMessageApplicationCommandInteractionDataResolved;
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-resolved-data-structure
+ */
+export interface APIMessageApplicationCommandInteractionDataResolved {
+	messages: Record<Snowflake, APIMessage>;
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-interaction-data-structure
+ */
+export type APIContextMenuInteractionData =
+	| APIUserApplicationCommandInteractionData
+	| APIMessageApplicationCommandInteractionData;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIUserApplicationCommandInteraction =
+	APIApplicationCommandInteractionWrapper<APIUserApplicationCommandInteractionData>;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIUserApplicationCommandDMInteraction = APIDMInteractionWrapper<APIUserApplicationCommandInteraction>;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIUserApplicationCommandGuildInteraction =
+	APIGuildInteractionWrapper<APIUserApplicationCommandInteraction>;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIMessageApplicationCommandInteraction =
+	APIApplicationCommandInteractionWrapper<APIMessageApplicationCommandInteractionData>;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIMessageApplicationCommandDMInteraction =
+	APIDMInteractionWrapper<APIMessageApplicationCommandInteraction>;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIMessageApplicationCommandGuildInteraction =
+	APIGuildInteractionWrapper<APIMessageApplicationCommandInteraction>;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIContextMenuInteraction = APIUserApplicationCommandInteraction | APIMessageApplicationCommandInteraction;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIContextMenuDMInteraction =
+	| APIUserApplicationCommandDMInteraction
+	| APIMessageApplicationCommandDMInteraction;
+
+/**
+ * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
+ */
+export type APIContextMenuGuildInteraction =
+	| APIUserApplicationCommandGuildInteraction
+	| APIMessageApplicationCommandGuildInteraction;
